Open the resource's package page in your browser

Opens a browser to the resource's package page on the City of Toronto Open Data Portal.

Usage

browse_resource(resource)

Arguments

resource

A way to identify the resource. Either a resource ID (passed as a character vector directly) or a single resource resulting from list_package_resources.

Value

A browser is opened to the resource's package page on the City of Toronto Open Data Portal website if the session is interactive. The URL is returned as a character string.

Examples

# \donttest{
ttc_subway_delays <- search_packages("ttc subway delay")
res <- list_package_resources(ttc_subway_delays)
browse_resource(res[1, ])
# }