Download a resource into your R session

Download a resource from the portal directly into your R session. CSV, XLS, XLSX, XML, JSON, SHP, ZIP, and GeoJSON resources are supported.

Usage

get_resource(resource)

Arguments

resource

A way to identify the resource. Either a resource ID (passed as a character vector directly) or a single resource resulting from list_package_resources.

Value

In most cases, the resource is returned as a tibble or list of tibbles. If it is a spatial resource (i.e. SHP or GeoJSON), it is returned as an sf object.

Examples

# \donttest{
list_package_resources("https://open.toronto.ca/dataset/neighbourhoods/") %>%
head(1) %>%
get_resource()
#> Simple feature collection with 158 features and 11 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -79.63926 ymin: 43.581 xmax: -79.11527 ymax: 43.85546
#> Geodetic CRS:  WGS 84
#> # A tibble: 158 × 12
#>    `_id` AREA_ID AREA_ATTR_ID PARENT_AREA_ID AREA_SHORT_CODE AREA_LONG_CODE
#>    <int>   <int>        <int> <lgl>          <chr>           <chr>         
#>  1     1 2502366     26022881 NA             174             174           
#>  2     2 2502365     26022880 NA             173             173           
#>  3     3 2502364     26022879 NA             172             172           
#>  4     4 2502363     26022878 NA             171             171           
#>  5     5 2502362     26022877 NA             170             170           
#>  6     6 2502361     26022876 NA             169             169           
#>  7     7 2502360     26022875 NA             156             156           
#>  8     8 2502359     26022874 NA             155             155           
#>  9     9 2502358     26022873 NA             154             154           
#> 10    10 2502357     26022872 NA             153             153           
#> # ℹ 148 more rows
#> # ℹ 6 more variables: AREA_NAME <chr>, AREA_DESC <chr>, CLASSIFICATION <chr>,
#> #   CLASSIFICATION_CODE <chr>, OBJECTID <int>, geometry <POLYGON [°]>
# }