Download a resource into your R session

Download a resource from the portal directly into your R session. CSV, XLS, XLSX, XML, JSON, SHP, ZIP, and GeoJSON resources are supported.

get_resource(resource)

Arguments

resource
A way to identify the resource. Either a resource ID (passed as a character vector directly) or a single resource resulting from list_package_resources.

Value

In most cases, the resource is returned as a tibble or list of tibbles. If it is a spatial resource (i.e. SHP or GeoJSON), it is returned as an sf object.

Examples

# \donttest{
res <- list_package_resources("1db34737-ffad-489d-a590-9171d500d453")
get_resource(resource = res[1, "id"])

#> Simple feature collection with 5738 features and 23 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -79.63465 ymin: 43.59329 xmax: -79.1231 ymax: 43.84012
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> # A tibble: 5,738 x 24
#>     `_id` OBJECTID ID    ADDRESSNUMBERTE… ADDRESSSTREET FRONTINGSTREET SIDE 
#>     <int>    <int> <chr> <chr>            <chr>         <chr>          <chr>
#>  1 621459        2 SH-0… 5810             Sheppard Ave… Sheppard Ave E North
#>  2 621460        6 SH-0… 4000             Jane St       Finch Ave W    North
#>  3 621461        8 SH-0… 1095             Neilson Rd    Neilson Rd     East 
#>  4 621462       12 SH-0… 47               Bathurst St   Bathurst St    East 
#>  5 621463       20 SH-0… 1524             Bathurst St   Bathurst St    West 
#>  6 621464       21 SH-0… 2345             Finch Ave W   Weston Rd      East 
#>  7 621465       23 SH-0… 6432             Finch Ave W   Finch Ave W    West 
#>  8 621466       26 SH-0… 755              York Mills Rd Leslie St      West 
#>  9 621467       27 SH-0… 122              Tapscott Rd   <NA>           <NA> 
#> 10 621468       28 SH-0… 338              Donlands Ave  Donlands Ave   West 
#> # … with 5,728 more rows, and 17 more variables: FROMSTREET <chr>,
#> #   DIRECTION <chr>, SITEID <chr>, WARD <chr>, BIA <chr>, ASSETTYPE <chr>,
#> #   STATUS <chr>, STYLE <chr>, BARCODE <chr>, POSTERBOARD <chr>,
#> #   POSTERBOARDSTATUS <chr>, SDE_STATE_ID <int>, X <dbl>, Y <dbl>,
#> #   LONGITUDE <dbl>, LATITUDE <dbl>, geometry <POINT [°]>

res[["id"]]

#> [1] "b9214fd7-60d1-45f3-8463-a6bd9828f8bf"

get_resource("b9214fd7-60d1-45f3-8463-a6bd9828f8bf")

#> Simple feature collection with 5738 features and 23 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -79.63465 ymin: 43.59329 xmax: -79.1231 ymax: 43.84012
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> # A tibble: 5,738 x 24
#>     `_id` OBJECTID ID    ADDRESSNUMBERTE… ADDRESSSTREET FRONTINGSTREET SIDE 
#>     <int>    <int> <chr> <chr>            <chr>         <chr>          <chr>
#>  1 621459        2 SH-0… 5810             Sheppard Ave… Sheppard Ave E North
#>  2 621460        6 SH-0… 4000             Jane St       Finch Ave W    North
#>  3 621461        8 SH-0… 1095             Neilson Rd    Neilson Rd     East 
#>  4 621462       12 SH-0… 47               Bathurst St   Bathurst St    East 
#>  5 621463       20 SH-0… 1524             Bathurst St   Bathurst St    West 
#>  6 621464       21 SH-0… 2345             Finch Ave W   Weston Rd      East 
#>  7 621465       23 SH-0… 6432             Finch Ave W   Finch Ave W    West 
#>  8 621466       26 SH-0… 755              York Mills Rd Leslie St      West 
#>  9 621467       27 SH-0… 122              Tapscott Rd   <NA>           <NA> 
#> 10 621468       28 SH-0… 338              Donlands Ave  Donlands Ave   West 
#> # … with 5,728 more rows, and 17 more variables: FROMSTREET <chr>,
#> #   DIRECTION <chr>, SITEID <chr>, WARD <chr>, BIA <chr>, ASSETTYPE <chr>,
#> #   STATUS <chr>, STYLE <chr>, BARCODE <chr>, POSTERBOARD <chr>,
#> #   POSTERBOARDSTATUS <chr>, SDE_STATE_ID <int>, X <dbl>, Y <dbl>,
#> #   LONGITUDE <dbl>, LATITUDE <dbl>, geometry <POINT [°]>

# }